+++ resources/win/cpm.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+"$(dirname "$0")/../app/cpm/bin/cpm" "$@"